requirements.txt
# Core FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# HTTP Client Libraries
httpx==0.24.1
requests==2.31.0

# AI and LangGraph Dependencies
openai==1.3.0
langgraph==0.0.62
langchain-core==0.1.52

# Database (if needed)
supabase==2.0.2

# Authentication & Security (Optional)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development Dependencies (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
